pyproject.toml
[project]
name = "audio-text-info-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "instructor>=1.13.0",
    "librosa>=0.11.0",
    "openai>=2.8.1",
    "openai-whisper>=20250625",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
    "torch>=2.9.1",
    "transformers>=4.57.3",
    "vosk>=0.3.45",
    "watchdog>=6.0.0",
]
